Reduce visibility of difficulty type inner values
